+++ mod/customcert/db/upgrade.php
@@ -135,7 +135,7 @@ function xmldb_customcert_upgrade($oldversion) {
         upgrade_mod_savepoint(true, 2017050506, 'customcert');
     }
 
-    if ($oldversion < 2017111306) {
+    if ($oldversion < 2018051705) {
         $table = new xmldb_table('customcert_elements');
         $field = new xmldb_field('element', XMLDB_TYPE_CHAR, '255', null, XMLDB_NOTNULL, null, null, 'name');
 
@@ -143,7 +143,7 @@ function xmldb_customcert_upgrade($oldversion) {
         $dbman->change_field_type($table, $field);
 
         // Savepoint reached.
-        upgrade_mod_savepoint(true, 2017111306, 'customcert');
+        upgrade_mod_savepoint(true, 2018051705, 'customcert');
     }
 
     return true;