+++ mod/customcert/tests/element_helper_test.php
@@ -194,13 +194,11 @@ class mod_customcert_element_helper_testcase extends advanced_testcase {
      * Test we return the correct grade information for an activity.
      */
     public function test_get_mod_grade_info() {
-        global $CFG;
-
-        // Set that we want 3 decimals to display.
-        $CFG->grade_decimalpoints = 3;
-
         // Create a course.
         $course = $this->getDataGenerator()->create_course();
+
+        // Set that we want 3 decimals to display.
+        grade_set_setting($course->id, 'decimalpoints', 3);
 
         // Create two users.
         $student1 = $this->getDataGenerator()->create_user();
@@ -257,22 +255,19 @@ class mod_customcert_element_helper_testcase extends advanced_testcase {
         $this->assertEquals(null, $grade->get_grade());
         $this->assertEquals('-', $grade->get_displaygrade());
         $this->assertEquals(null, $grade->get_dategraded());
+
+        grade_get_setting($course->id, null, null, true);
     }
 
     /**
      * Test we return the correct grade information for a course.
      */
     public function test_get_course_grade_info() {
-        global $CFG;
-
-        // Including to use constant.
-        require_once($CFG->dirroot . '/mod/customcert/element/grade/classes/element.php');
-
-        // Set that we want 3 decimals to display.
-        $CFG->grade_decimalpoints = 3;
-
         // Create a course.
         $course = $this->getDataGenerator()->create_course();
+
+        // Set that we want 3 decimals to display.
+        grade_set_setting($course->id, 'decimalpoints', 3);
 
         // Create two users.
         $student1 = $this->getDataGenerator()->create_user();
@@ -320,19 +315,19 @@ class mod_customcert_element_helper_testcase extends advanced_testcase {
         $this->assertEquals(null, $grade->get_grade());
         $this->assertEquals('-', $grade->get_displaygrade());
         $this->assertEquals(null, $grade->get_dategraded());
+
+        grade_get_setting($course->id, null, null, true);
     }
 
     /**
      * Test we return the correct grade information for a grade item.
      */
     public function test_get_grade_item_info() {
-        global $CFG;
-
-        // Set that we want 3 decimals to display.
-        $CFG->grade_decimalpoints = 3;
-
         // Create a course.
         $course = $this->getDataGenerator()->create_course();
+
+        // Set that we want 3 decimals to display.
+        grade_set_setting($course->id, 'decimalpoints', 3);
 
         // Create two users.
         $student1 = $this->getDataGenerator()->create_user();
@@ -382,5 +377,7 @@ class mod_customcert_element_helper_testcase extends advanced_testcase {
         $this->assertEquals(null, $grade->get_grade());
         $this->assertEquals('-', $grade->get_displaygrade());
         $this->assertEquals(null, $grade->get_dategraded());
+
+        grade_get_setting($course->id, null, null, true);
     }
 }