+++ mod/customcert/tests/email_certificate_task_test.php
@@ -45,6 +45,39 @@ class mod_customcert_task_email_certificate_task_testcase extends advanced_testc
     }
 
     /**
+     * Tests the email certificate task for users without a capability to receive a certificate.
+     */
+    public function test_email_certificates_no_cap() {
+        global $DB;
+
+        // Create a course.
+        $course = $this->getDataGenerator()->create_course();
+
+        // Create some users.
+        $user1 = $this->getDataGenerator()->create_user();
+        $user2 = $this->getDataGenerator()->create_user();
+
+        // Enrol two of them in the course as students but revoke their right to receive a certificate issue.
+        $roleids = $DB->get_records_menu('role', null, '', 'shortname, id');
+        $this->getDataGenerator()->enrol_user($user1->id, $course->id);
+        $this->getDataGenerator()->enrol_user($user2->id, $course->id);
+
+        unassign_capability('mod/customcert:receiveissue', $roleids['student']);
+
+        // Create a custom certificate.
+        $this->getDataGenerator()->create_module('customcert', ['course' => $course->id, 'emailstudents' => 1]);
+
+        // Run the task.
+        $sink = $this->redirectEmails();
+        $task = new \mod_customcert\task\email_certificate_task();
+        $task->execute();
+        $emails = $sink->get_messages();
+
+        // Confirm that we did not send any emails.
+        $this->assertCount(0, $emails);
+    }
+
+    /**
      * Tests the email certificate task for students.
      */
     public function test_email_certificates_students() {
@@ -95,11 +128,9 @@ class mod_customcert_task_email_certificate_task_testcase extends advanced_testc
         // Confirm that we sent out emails to the two users.
         $this->assertCount(2, $emails);
 
-        $this->assertContains(fullname($user3), $emails[0]->header);
         $this->assertEquals($CFG->noreplyaddress, $emails[0]->from);
         $this->assertEquals($user1->email, $emails[0]->to);
 
-        $this->assertContains(fullname($user3), $emails[1]->header);
         $this->assertEquals($CFG->noreplyaddress, $emails[1]->from);
         $this->assertEquals($user2->email, $emails[1]->to);
 
@@ -150,11 +181,9 @@ class mod_customcert_task_email_certificate_task_testcase extends advanced_testc
         // Confirm that we only sent out 2 emails, both emails to the teacher for the two students.
         $this->assertCount(2, $emails);
 
-        $this->assertContains(fullname($user3), utf8_encode($emails[0]->header));
         $this->assertEquals($CFG->noreplyaddress, $emails[0]->from);
         $this->assertEquals($user3->email, $emails[0]->to);
 
-        $this->assertContains(fullname($user3), utf8_encode($emails[1]->header));
         $this->assertEquals($CFG->noreplyaddress, $emails[1]->from);
         $this->assertEquals($user3->email, $emails[1]->to);
     }
@@ -189,11 +218,9 @@ class mod_customcert_task_email_certificate_task_testcase extends advanced_testc
         // Confirm that we only sent out 2 emails, both emails to the other address that was valid for the two students.
         $this->assertCount(2, $emails);
 
-        $this->assertContains(fullname(get_admin()), utf8_encode($emails[0]->header));
         $this->assertEquals($CFG->noreplyaddress, $emails[0]->from);
         $this->assertEquals('testcustomcert@example.com', $emails[0]->to);
 
-        $this->assertContains(fullname(get_admin()), utf8_encode($emails[1]->header));
         $this->assertEquals($CFG->noreplyaddress, $emails[1]->from);
         $this->assertEquals('testcustomcert@example.com', $emails[1]->to);
     }